:bug: fix(neovim/p): Remove buffer autocompletion
Not sure why I set this up when I  could  easily  use  Neovim's  builtin
buffer completion (<C-x><C-[np]>), which uses less memory and is  faster
anyway. Granted, the UI is not as unified, but  to  hell  with  it;  I'm
tired of having Neovim cold starting at +100M of memory.