refactor(neovim/p): Enhance & reduce `nvim-cmp`
This is a long overdue overhaul of `nvim-cmp`'s configuration, disabling
most of its functionality and keeping the bare minimum.

This is because using this plugin for a long time felt like a hindrance
more than a helper. Especially the `performance` part is broken at best
and freezes the entire neovim instance when throttling.