neovim/p: Overhaul `telescope.nvim` mappings
The old (f)ind mnemonic was weird in that it requires some finger
gymnastics to work on Dvorak. The new (s)earch mnemonic feel more
natural and is more finger-friendly. The `s` key feels a bit
overburdened with mappings, but it works for now.

I'm now also avoiding double-tap because it's slower and breaks my flow.