neovim/sn: Add some snippets
Some shortcuts for commonly written pieces of code. These are slow to
write manually, therefore it makes sense to add as snippets.

For TypeScript, I use the JSDoc comment a lot and writing it manually
each time is a pain.

Same for shell scripting, the `then` keyword followed by a newline, then
`fi` while triggering the LSP is annoying.

Signed-Off-By: scheatkode <scheatkode@disroot.org>